app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.28.1
aiogram==3.15.0
python-multipart==0.0.20
jinja2==3.1.4
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
openpyxl==3.1.5
pandas==2.2.3
lxml==5.3.0
python-dotenv==1.0.1
